src/use-cases/test/fetch-user-check-ins-history.spec.ts
import { expect, describe, it, beforeEach, vi, afterEach } from 'vitest'
import { InMemoryCheckinsRepository } from '@/repositories/in-memory/in-memory-check-ins-repository'
import { FetchUserCheckInHistoryUseCase } from '../fetch-user-check-ins-history'


let checkInRepository: InMemoryCheckinsRepository
let sut: FetchUserCheckInHistoryUseCase


//Testes no caso de uso dos chekins
describe('CheckIn use cases', () => {
   beforeEach(async () => {
      checkInRepository = new InMemoryCheckinsRepository()
      sut = new FetchUserCheckInHistoryUseCase(checkInRepository)
   })

   // Deve ser possivel obter a lista de checkins de um usuário
   it('should be able to fetch Check-in history', async () => {

      await checkInRepository.create({
         gym_id: 'gym-01',
         user_id: 'user-01'
      })
      await checkInRepository.create({
         gym_id: 'gym-02',
         user_id: 'user-01'
      })

      const { checkIns } = await sut.execute({ userId: 'user-01', page: 1 })

      expect(checkIns).toHaveLength(2)
      expect(checkIns).toEqual([
         expect.objectContaining({ gym_id: 'gym-01' }),
         expect.objectContaining({ gym_id: 'gym-02' })
      ])
   })

   it('should be able to fetch paginated Check-in history', async () => {

      for (let i = 1; i <= 22; i++) {
         await checkInRepository.create({
            gym_id: `gym-${i}`,
            user_id: 'user-01'
         })
      }

      const { checkIns } = await sut.execute({ userId: 'user-01', page: 2 })

      expect(checkIns).toEqual([
         expect.objectContaining({ gym_id: 'gym-21' }),
         expect.objectContaining({ gym_id: 'gym-22' })
      ])
   })

})